feat: remove benchmarking, strenghten test processes
