--- docs/contribute/benchmarks.rst
@@ -1,10 +1,0 @@
-Benchmarks
-==========
-
-https://pytest-benchmark.readthedocs.io/
-
-
-Plotting
-::::::::
-
-pytest-benchmark[histogram]
