+++ harp/utils/testing/__init__.py
@@ -3,7 +3,7 @@ import re
 import pytest
 
 pytest.register_assert_rewrite("harp.utils.testing.applications")
-pytest.register_assert_rewrite("harp.utils.testing.benchmarking")
+pytest.register_assert_rewrite("harp.utils.testing.cli")
 pytest.register_assert_rewrite("harp.utils.testing.config")
