+++ pyproject.toml
@@ -66,7 +66,6 @@ dev = [
     "pre-commit>=3.7.1,<5.0.0",
     "pytest>=7.4.3,<9.0.0",
     "pytest-asyncio>=0.26.0,<0.27",
-    "pytest-benchmark[histogram]>=4,<6",
     "pytest-cov>=4.1,<7.0",
     "pytest-timeout>=2.3.1,<3",
     "pytest-xdist[psutil]>=3.6.1,<4",
